Add dynamic sidebar component
